fixing the namespace Fixed the namespace, also added a new inline function.
